# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API для управления продуктами, категориями и брендами.",
        "title": "Product Service API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8082",
    "basePath": "/",
    "paths": {
...
        "/products": {
            "get": {
                "description": "Возвращает список всех продуктов",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Продукты"
                ],
                "summary": "Получение всех продуктов",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/internal_product.Product"
                            }
                        }
                    },
                    "500": {
                        "description": "Ошибка получения продуктов",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Создаёт новый продукт с переданными данными",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Продукты"
                ],
                "summary": "Создание нового продукта",
                "parameters": [
                    {
                        "description": "Данные продукта",
                        "name": "product",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_product.Product"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/internal_product.Product"
                        }
                    },
                    "400": {
                        "description": "Неверное тело запроса",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Ошибка при создании продукта",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/products/questions/{id}": {
            "get": {
                "description": "Возвращает список вопросов для заданного варианта продукта",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Вопросы"
                ],
                "summary": "Получение вопросов по ID варианта продукта",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID варианта продукта",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Количество вопросов для получения",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Смещение для пагинации",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/service.QuestionListResponse"
                        }
                    },
                    "400": {
                        "description": "Неверный ID продукта",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Ошибка получения вопросов",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/products/reviews/{id}": {
            "get": {
                "description": "Возвращает список отзывов для заданного варианта продукта",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Отзывы"
                ],
                "summary": "Получение отзывов по ID варианта продукта",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID варианта продукта",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Количество отзывов для получения",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Смещение для пагинации",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/service.ReviewListResponse"
                        }
                    },
                    "400": {
                        "description": "Неверный ID продукта",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Ошибка получения отзывов",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_ShopOnGO_product-service_internal_brand.Brand": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "logo": {
                    "description": "JSON хранящий ссылку на статику(изображение)",
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "video_url": {
                    "description": "Ссылка на видео в облаке",
                    "type": "string"
                }
            }
        },
        "github_com_ShopOnGO_product-service_internal_category.Category": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "image_url": {
                    "description": "Ссылка на изображение категории",
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "parentCategory": {
                    "$ref": "#/definitions/github_com_ShopOnGO_product-service_internal_category.Category"
                },
                "parentCategoryID": {
                    "description": "Внешний ключ может быть NULL",
                    "type": "integer"
                },
                "subCategories": {
                    "description": "Связь для подкатегорий",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_ShopOnGO_product-service_internal_category.Category"
                    }
                }
            }
        },
        "github_com_ShopOnGO_product-service_internal_productVariant.ProductVariant": {
            "type": "object",
            "properties": {
                "barcode": {
                    "description": "Weight          uint      \t\t`gorm:\"default:0\"`         // Вес в граммах",
                    "type": "string"
                },
                "colors": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "dimensions": {
                    "description": "Габариты (например \"20x30x5 см\")",
                    "type": "string"
                },
                "discount": {
                    "type": "number"
                },
                "id": {
                    "type": "integer"
                },
                "imageURLs": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "isActive": {
                    "description": "Активен ли вариант",
                    "type": "boolean"
                },
                "minOrder": {
                    "description": "Минимальный заказ",
                    "type": "integer"
                },
                "price": {
                    "type": "number"
                },
                "productID": {
                    "description": "на всякий",
                    "type": "integer"
                },
                "rating": {
                    "type": "number"
                },
                "ratingSum": {
                    "type": "integer"
                },
                "reservedStock": {
                    "description": "бронь (пока оплатишь типа)",
                    "type": "integer"
                },
                "reviewCount": {
                    "type": "integer"
                },
                "sizes": {
                    "type": "string"
                },
                "sku": {
                    "description": "Уникальный артикул",
                    "type": "string"
                },
                "stock": {
                    "description": "Общий остаток на складе",
                    "type": "integer"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "gorm.DeletedAt": {
            "type": "object",
            "properties": {
                "time": {
                    "type": "string"
                },
                "valid": {
                    "description": "Valid is true if Time is not NULL",
                    "type": "boolean"
                }
            }
        },
...
        "internal_product.Product": {
            "type": "object",
            "properties": {
                "brand": {
                    "$ref": "#/definitions/github_com_ShopOnGO_product-service_internal_brand.Brand"
                },
                "brand_id": {
                    "type": "integer"
                },
                "category": {
                    "$ref": "#/definitions/github_com_ShopOnGO_product-service_internal_category.Category"
                },
                "category_id": {
                    "type": "integer"
                },
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "imageURLs": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "is_active": {
                    "type": "boolean"
                },
                "material": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "variants": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_ShopOnGO_product-service_internal_productVariant.ProductVariant"
                    }
                },
                "videoURLs": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
...
        "service.Model": {
            "type": "object",
            "properties": {
                "created_at": {
                    "$ref": "#/definitions/timestamppb.Timestamp"
                },
                "deleted_at": {
                    "$ref": "#/definitions/timestamppb.Timestamp"
                },
                "id": {
                    "type": "integer"
                },
                "updated_at": {
                    "$ref": "#/definitions/timestamppb.Timestamp"
                }
            }
        },
        "service.Question": {
            "type": "object",
            "properties": {
                "answer_text": {
                    "type": "string"
                },
                "author": {
                    "description": "Types that are valid to be assigned to Author:\n\n\t*Question_UserId\n\t*Question_GuestId"
                },
                "likes_count": {
                    "type": "integer"
                },
                "model": {
                    "$ref": "#/definitions/service.Model"
                },
                "product_variant_id": {
                    "type": "integer"
                },
                "question_text": {
                    "type": "string"
                }
            }
        },
        "service.QuestionListResponse": {
            "type": "object",
            "properties": {
                "questions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/service.Question"
                    }
                }
            }
        },
        "service.Review": {
            "type": "object",
            "properties": {
                "comment": {
                    "type": "string"
                },
                "likes_count": {
                    "type": "integer"
                },
                "model": {
                    "$ref": "#/definitions/service.Model"
                },
                "product_variant_id": {
                    "type": "integer"
                },
                "rating": {
                    "type": "integer"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "service.ReviewListResponse": {
            "type": "object",
            "properties": {
                "reviews": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/service.Review"
                    }
                }
            }
        },
        "timestamppb.Timestamp": {
            "type": "object",
            "properties": {
                "nanos": {
                    "description": "Non-negative fractions of a second at nanosecond resolution. Negative\nsecond values with fractions must still have non-negative nanos values\nthat count forward in time. Must be from 0 to 999,999,999\ninclusive.",
                    "type": "integer"
                },
                "seconds": {
                    "description": "Represents seconds of UTC time since Unix epoch\n1970-01-01T00:00:00Z. Must be from 0001-01-01T00:00:00Z to\n9999-12-31T23:59:59Z inclusive.",
                    "type": "integer"
                }
            }
        }
    }
}